linting and file reorganisation
