+++ eslint.config.mjs
@@ -1,0 +1,14 @@
+import globals from "globals";
+import pluginJs from "@eslint/js";
+import tseslint from "typescript-eslint";
+
+
+/** @type {import('eslint').Linter.Config[]} */
+export default [
+  {files: ["**/*.{js,mjs,cjs,ts}"]},
+  {files: ["**/*.js"], languageOptions: {sourceType: "script"}},
+  {languageOptions: { globals: globals.browser }},
+  {rules: {"@typescript-eslint/no-explicit-any": "off"}},
+  pluginJs.configs.recommended,
+  ...tseslint.configs.recommended,
+];
