+++ package.json
@@ -5,6 +5,7 @@
   "main": "dist/index.js",
   "scripts": {
     "build": "tsc",
+    "lint": "npx eslint ./src",
     "postbuild": "chmod +x dist/index.js",
     "start": "ts-node src/index.ts",
     "blessdev": "nodemon -I --exec ts-node src/test.ts",
@@ -17,12 +18,19 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@eslint/js": "^9.22.0",
     "@types/blessed": "^0.1.25",
     "@types/fs-extra": "^11.0.4",
+    "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.10.1",
     "@types/ws": "^8.5.13",
+    "@typescript-eslint/eslint-plugin": "^8.26.1",
+    "@typescript-eslint/parser": "^8.26.1",
+    "eslint": "^9.22.0",
+    "globals": "^16.0.0",
     "ts-node": "^10.9.2",
-    "typescript": "^5.7.2"
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.26.1"
   },
   "dependencies": {
     "@types/inquirer": "^9.0.7",
