--- src/services/configs/index.ts
@@ -1,8 +1,0 @@
-export * from './get-participants-config';
-export * from './get-l1-config';
-export * from './get-signer-config';
-export * from './get-chain-config';
-export * from './get-gas-config';
-export * from './get-da-config';
-export * from './get-interop-config';
-export * from './confirm-config';
